site: use theme-generator-plugin (#496)
* feat: replace color with theme generator plugin

* feat: remove variables

* feat: upgrade v1

* chore: upgrade v1

* feat: adjust page config panel style

* feat: add dock

* chore: remove dead code

* feat: upgrade setting config

* feat: complete page config detail

* feat: upgrade theme generator
