+++ package.json
@@ -29,7 +29,8 @@
     "qrcode.vue": "^3.3.4",
     "qs": "^6.11.1",
     "tdesign-icons-vue-next": "^0.1.8",
-    "tdesign-vue-next": "^1.2.2",
+    "tdesign-theme-generator": "^1.0.4",
+    "tdesign-vue-next": "^1.3.1",
     "tvision-color": "^1.5.0",
     "vue": "^3.2.47",
     "vue-clipboard3": "^2.0.0",
