+++ src/App.vue
@@ -1,9 +1,11 @@
 <template>
   <router-view :class="[mode]" />
+  <setting-component />
 </template>
 <script setup lang="ts">
 import { computed } from 'vue';
 
+import SettingComponent from '@/layouts/setting/index.vue';
 import { useSettingStore } from '@/store';
 
 const store = useSettingStore();
