+++ src/main.ts
@@ -10,6 +10,8 @@ import 'tdesign-vue-next/es/style/index.css';
 import '@/style/index.less';
 import './permission';
 
+import 'tdesign-theme-generator';
+
 const app = createApp(App);
 
 app.use(TDesign);
