+++ src/pages/login/index.vue
@@ -16,7 +16,6 @@
 
       <login v-if="type === 'login'" />
       <register v-else @register-success="switchType('login')" />
-      <tdesign-setting />
     </div>
 
     <footer class="copyright">Copyright @ 2021-2022 Tencent. All Rights Reserved</footer>
@@ -29,8 +28,6 @@ export default {
 </script>
 <script setup lang="ts">
 import { ref } from 'vue';
-
-import TdesignSetting from '@/layouts/setting.vue';
 
 import LoginHeader from './components/Header.vue';
 import Login from './components/Login.vue';
